add cache expiration and flush on shutdown
This addresses two concerns:
1. since there's no way for logspout to signal the adapter that a
container has stopped, we depend on a time signal to garbage collect old
buffers
2. We must buffer each line because the next line might be part of a
continuation.  Therefor, we must wait at least until a new
line is written before delivering the buffered line.  Cache expiration
provides another signal for flushing buffered lines, without which we
would never write the last line a container logged.  This also puts an
upper bound on the amount of time the last log line is buffered.

Implementation Note:
We do not do an expiration check on each message streamed.  Instead, we
setup a tick channel that fires at the rate of cache TTL.  Receipt of
that signal indicates a cache expiration is needed.  The same go thread
that manages the cache must handle this signal since the cache is not
thread safe.